requirements.txt
celery
flower
mysql-replication
pymysql
redis
streamlit
zodb
